refactor: use shared layout
