+++ apps/dashboard/src/app/my-images/page.tsx
@@ -1,5 +1,4 @@
-import { Suspense } from "react";
-import { OgSplash } from "../../components/OgSplash";
+import { MyImagesSplash } from "../../components/Splash/MyImagesSplash";
 
 export const metadata = {
   title: 'My images - OG Studio',
@@ -7,11 +6,5 @@ export const metadata = {
 }
 
 export default function MyImages() {
-  return (
-    // OgSplash uses `useSearchParams()` so we need to wrap it in a Suspense
-    // to allow to statically render the page: https://nextjs.org/docs/app/building-your-application/rendering/server-components#dynamic-functions
-    <Suspense>
-      <OgSplash route="my-images" />
-    </Suspense>
-  )
+  return <MyImagesSplash />
 }
