+++ apps/dashboard/src/app/templates/page.tsx
@@ -1,5 +1,4 @@
-import { Suspense } from "react";
-import { OgSplash } from "../../components/OgSplash";
+import { TemplatesSplash } from "../../components/Splash/TemplatesSplash";
 
 export const metadata = {
   title: 'Templates - OG Studio',
@@ -7,11 +6,5 @@ export const metadata = {
 }
 
 export default function Templates() {
-  return (
-    // OgSplash uses `useSearchParams()` so we need to wrap it in a Suspense
-    // to allow to statically render the page: https://nextjs.org/docs/app/building-your-application/rendering/server-components#dynamic-functions
-    <Suspense>
-      <OgSplash route="templates" />
-    </Suspense>
-  )
+  return <TemplatesSplash />
 }
